requirements.txt
google-cloud-translate 